+++ src/image_loader/image_loader_stb.c
@@ -40,11 +40,11 @@ ret_t stb_load_image(const uint8_t* buff, uint32_t buff_size, bitmap_t* image, b
   return_value_if_fail(data != NULL, RET_FAIL);
 
   if (enable_bgr565 && rgba_data_is_opaque(data, w, h, n)) {
-    ret = bitmap_init(image, w, h, BITMAP_FMT_BGR565, data, n);
+    ret = bitmap_init_from_rgba(image, w, h, BITMAP_FMT_BGR565, data, n);
   } else if (require_bgra) {
-    ret = bitmap_init(image, w, h, BITMAP_FMT_BGRA8888, data, n);
+    ret = bitmap_init_from_rgba(image, w, h, BITMAP_FMT_BGRA8888, data, n);
   } else {
-    ret = bitmap_init(image, w, h, BITMAP_FMT_RGBA8888, data, n);
+    ret = bitmap_init_from_rgba(image, w, h, BITMAP_FMT_RGBA8888, data, n);
   }
 
   stbi_image_free((uint8_t*)(data));